+++ recipes/schedule_run.rb
@@ -5,7 +5,7 @@
 # Copyright:: 2017, The Authors, All Rights Reserved.
 
 ### Add a crontab entry to run laravel artisan scheduler
-artisan 'schedule:run >> /dev/null 2>&1' do
+artisan_schedule 'schedule:run >> /dev/null 2>&1' do
 	path node['laravel-artisan']['path']
 	action :create
 end