+++ recipes/schedule_stop.rb
@@ -5,7 +5,7 @@
 # Copyright:: 2017, The Authors, All Rights Reserved.
 
 ### Remove the crontab entry for the scheduler
-artisan 'stop scheduler' do
+artisan_schedule 'stop scheduler' do
 	path node['laravel-artisan']['path']
-	action :remove
+	action :delete
 end